v2.7: notebook 改相对路径 + 新基础数据全量重训
- 01/02/03 cells.txt 去掉 d:/Qunar_work 绝对路径，改为代码目录上一级的相对路径（兼容 LEIDEN_BASE 环境变量）
- 输入切换为 26.08.27_base.csv（735K 设备，重导出修复 device_id 科学计数法）
- 01 修复新数据缺 flight_pay_tool_detail 列的 KeyError（演示 cell 加存在性判断）
- server 修复 community_id '6.0' 字符串转 int 崩溃（float 中转）
- 全量重训：02 八分半 / 01 十三分钟 / 03 合并全部通过，server 已加载新数据
- 新数据结果：高风险 20949(2.9%) / 中风险 202536(27.6%) / 疑似 469953(63.9%) / 普通 41518(5.7%)；社区 19194 个，图节点 314K

+++ tools/community_server.py
@@ -70,7 +70,8 @@ class GraphEngine:
         comm_df = pd.read_csv(comm_path, encoding="utf-8-sig", dtype=str)
         # 过滤掉 NaN community_id (被筛除的小社区)
         comm_df = comm_df[comm_df["community_id"].notna()].copy()
-        comm_df["community_id"] = comm_df["community_id"].astype(int)
+        # 先转 float 再转 int，兼容 "6.0" 这类字符串（dtype=str 读入时常见）
+        comm_df["community_id"] = comm_df["community_id"].astype(float).astype(int)
         for _, row in tqdm(comm_df.iterrows(), total=len(comm_df), desc="  社区映射"):
             self.node_type[row["node"]] = row["node_type"]
             self.node_comm[row["node"]] = int(row["community_id"])
@@ -410,7 +411,8 @@ class GraphEngine:
         # 合并团伙高危标记 (来自 gang_list.csv)
         if self.gang_df is not None and "is_high_risk_gang" in self.gang_df.columns:
             gang_info = self.gang_df[["community_id", "is_high_risk_gang"]].copy()
-            gang_info["community_id"] = gang_info["community_id"].astype(int)
+            # float 中转兼容 "6.0" 字符串
+            gang_info["community_id"] = gang_info["community_id"].astype(float).astype(int)
             result = result.merge(gang_info, on="community_id", how="left")
             result["is_high_risk_gang"] = result["is_high_risk_gang"].fillna(0).astype(int)
        # 排序
